+++ webapp/src/components/icons/slide-layout-no-image-left-align.tsx
@@ -11,8 +11,8 @@ export function SlideLayoutNoImageLeftAlign(props: React.SVGAttributes<any>) {
             {...props}
         >
             <rect x="0.5" y="0.5" width="81" height="50" rx="8.5" fill="white" />
-            <rect x="11" y="17" width="15" height="4" rx="2" fill="#DBDBDB" />
-            <rect x="11" y="25" width="30" height="8" rx="2" fill="#DBDBDB" />
+            <rect x="11" y="17" width="15" height="4" rx="2" fill="#3A465A" />
+            <rect x="11" y="25" width="30" height="8" rx="2" fill="white" stroke="#CBD5E6" />
         </svg>
     );
 }